fix: 修复第二轮 Kotlin 编译错误
- ByteQueue.kt: 使用 synchronized(this) + (this as java.lang.Object).notify/wait() 替代 @Synchronized 注解
- RustEngineCallback.kt: 添加 reportTitleChange/onClearScreen 的 override 修饰符,
  修复 onColorsChanged/onBell 便捷方法参数传递,
  修复 ByteArray.isNotEmpty() 替代 isNullOrEmpty(),
  添加 onSixelImage 到 TerminalSessionClient 接口
- TerminalSession.kt: 移除 getTitle 的 override (TerminalOutput 没有该方法),
  修复 cwd 空值处理 (cwd ?: ""), 修复 Throwable->Exception 类型转换,
  修复 onCopyTextToClipboard 空值安全, 初始化 bytesRead = 0
- TextStyle.kt: 0xff000000 添加 .toInt() 后缀解决 Int/Long 比较问题
- WcWidth.kt: @Deprecated 添加 message 参数

+++ terminal-emulator/src/main/java/com/termux/terminal/TextStyle.kt
@@ -32,12 +32,12 @@ object TextStyle {
     @JvmStatic
     fun encode(foreColor: Int, backColor: Int, effect: Int): Long {
         var result = (effect and 0b11111111111).toLong()
-        if (foreColor and 0xff000000 == 0xff000000) {
+        if (foreColor and 0xff000000.toInt() == 0xff000000.toInt()) {
             result = result or (CHARACTER_ATTRIBUTE_TRUECOLOR_FOREGROUND.toLong() or ((foreColor.toLong() and 0x00ffffffL) shl 40))
         } else {
             result = result or ((foreColor.toLong() and 0b111111111L) shl 40)
         }
-        if (backColor and 0xff000000 == 0xff000000) {
+        if (backColor and 0xff000000.toInt() == 0xff000000.toInt()) {
             result = result or (CHARACTER_ATTRIBUTE_TRUECOLOR_BACKGROUND.toLong() or ((backColor.toLong() and 0x00ffffffL) shl 16))
         } else {
             result = result or ((backColor.toLong() and 0b111111111L) shl 16)
